+++ contracts/protocol/pool/Pool.sol
@@ -704,18 +704,18 @@ contract Pool is VersionedInitializable, IPool, PoolStorage {
   }
 
   function _addReserveToList(address asset) internal virtual {
-    uint256 reservesCount = _reservesCount;
+    uint16 reservesCount = _reservesCount;
 
     require(reservesCount < MAX_NUMBER_RESERVES(), Errors.P_NO_MORE_RESERVES_ALLOWED);
 
     bool reserveAlreadyAdded = _reserves[asset].id != 0 || _reservesList[0] == asset;
 
     if (!reserveAlreadyAdded) {
-      for (uint8 i = 0; i <= reservesCount; i++) {
+      for (uint16 i = 0; i <= reservesCount; i++) {
         if (_reservesList[i] == address(0)) {
           _reserves[asset].id = i;
           _reservesList[i] = asset;
-          _reservesCount = uint16(reservesCount + 1);
+          _reservesCount = reservesCount + 1;
         }
       }
     }
